refactor(config): break the registry/parser import cycle with an identity leaf
models.py built AGENT_REGISTRY by calling the catalog parser, and the
parser imported the registry dataclasses back from models.py. Deferring
that import to call time made it run, but it is still a cycle — to the
[no_circular_imports] hard convention check, which walks every import in
the AST, and to anyone reading it.

Extract the identity half into config/model_identity.py: TransportSpec,
RoutingPolicy, AgentSpec, transport_for and the runner tables, the
canonical-id helpers, the tier-derived routing defaults, and the
raw-input alias boundary. It depends only on config/pricing.py (stdlib
only), so both model_catalog.py and models.py sit on it and neither
imports the other. models.py re-exports the whole surface, so every
existing `from theforge.config.models import ...` is unaffected.

models.py loads the catalog at module scope again (no deferred import),
and drops the eight _TRANSPORT_* constants whose comment claimed the
registry entries referenced them — nothing has referenced them for some
time, and they sat inside the moved block.

Adds layering tests so the back-edge cannot return silently, and records
the intended direction in config/CONVENTIONS.md.
